fix(dashboard): Add service chart
